Reimplemented the file descriptor handling and inter-thread communcation using epoll()
Removed libev dependency for libmetha. Implemented a new file descriptor listening system using epoll(), significantly faster inter-thread communication.

cleaned up lots of warnings messages

Added the LMC_DIRECTIVE() macro to lmc

+++ src/libmetha/umex.c
@@ -551,7 +551,7 @@ umex_dump(EXPR* expr)
 {
     printf("begin dump\n");
     printf("  (struct umex*) (%p)\n", expr);
-    printf("  %d bytes\n", sizeof(EXPR));
+    printf("  %d bytes\n", (int)sizeof(EXPR));
     printf("unsigned int\n");
     printf("  %d\n", expr->sz);
     printf("code\n");
